Restructure project
