+++ java/src/Utility/CustomLogger.java
@@ -1,4 +1,4 @@
-package Utility;
+package src.Utility;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
